Refactor AWS db mocks (#30086)
* Refactor AWS db mocks

* removew lib/services dependency

* fix rds instance url, add elasticache reader endpoint
