+++ lib/services/database.go
@@ -793,6 +793,60 @@ func NewDatabasesFromRDSClusterCustomEndpoints(cluster *rds.DBCluster) (types.Da
 	return databases, trace.NewAggregate(errors...)
 }
 
+// NewDatabasesFromRDSCluster creates all database resources from an RDS Aurora
+// cluster.
+func NewDatabasesFromRDSCluster(cluster *rds.DBCluster) (types.Databases, error) {
+	var errors []error
+	var databases types.Databases
+
+	// Find out what types of instances the cluster has. Some examples:
+	// - Aurora cluster with one instance: one writer
+	// - Aurora cluster with three instances: one writer and two readers
+	// - Secondary cluster of a global database: one or more readers
+	var hasWriterInstance, hasReaderInstance bool
+	for _, clusterMember := range cluster.DBClusterMembers {
+		if clusterMember != nil {
+			if aws.BoolValue(clusterMember.IsClusterWriter) {
+				hasWriterInstance = true
+			} else {
+				hasReaderInstance = true
+			}
+		}
+	}
+
+	// Add a database from primary endpoint, if any writer instances.
+	if cluster.Endpoint != nil && hasWriterInstance {
+		database, err := NewDatabaseFromRDSCluster(cluster)
+		if err != nil {
+			errors = append(errors, err)
+		} else {
+			databases = append(databases, database)
+		}
+	}
+
+	// Add a database from reader endpoint, if any reader instances.
+	// https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/Aurora.Overview.Endpoints.html#Aurora.Endpoints.Reader
+	if cluster.ReaderEndpoint != nil && hasReaderInstance {
+		database, err := NewDatabaseFromRDSClusterReaderEndpoint(cluster)
+		if err != nil {
+			errors = append(errors, err)
+		} else {
+			databases = append(databases, database)
+		}
+	}
+
+	// Add databases from custom endpoints
+	if len(cluster.CustomEndpoints) > 0 {
+		customEndpointDatabases, err := NewDatabasesFromRDSClusterCustomEndpoints(cluster)
+		if err != nil {
+			errors = append(errors, err)
+		}
+		databases = append(databases, customEndpointDatabases...)
+	}
+
+	return databases, trace.NewAggregate(errors...)
+}
+
 // NewDatabaseFromRDSProxy creates database resource from RDS Proxy.
 func NewDatabaseFromRDSProxy(dbProxy *rds.DBProxy, port int64, tags []*rds.Tag) (types.Database, error) {
 	metadata, err := MetadataFromRDSProxy(dbProxy)
@@ -908,6 +962,27 @@ func NewDatabasesFromElastiCacheNodeGroups(cluster *elasticache.ReplicationGroup
 		}
 	}
 	return databases, nil
+}
+
+// NewDatabasesFromElastiCacheReplicationGroup creates all database resources
+// from an ElastiCache ReplicationGroup.
+func NewDatabasesFromElastiCacheReplicationGroup(cluster *elasticache.ReplicationGroup, extraLabels map[string]string) (types.Databases, error) {
+	// Create database using configuration endpoint for Redis with cluster
+	// mode enabled.
+	if aws.BoolValue(cluster.ClusterEnabled) {
+		database, err := NewDatabaseFromElastiCacheConfigurationEndpoint(cluster, extraLabels)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		return types.Databases{database}, nil
+	}
+
+	// Create databases using primary and reader endpoints for Redis with
+	// cluster mode disabled. When cluster mode is disabled, it is expected
+	// there is only one node group (aka shard) with one primary endpoint
+	// and one reader endpoint.
+	databases, err := NewDatabasesFromElastiCacheNodeGroups(cluster, extraLabels)
+	return databases, trace.Wrap(err)
 }
 
 // newElastiCacheDatabase returns a new ElastiCache database.
