+++ lib/srv/discovery/fetchers/db/aws_elasticache_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package db
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -33,8 +32,8 @@ import (
 func TestElastiCacheFetcher(t *testing.T) {
 	t.Parallel()
 
-	elasticacheProd, elasticacheDatabaseProd, elasticacheProdTags := makeElastiCacheCluster(t, "ec1", "us-east-1", "prod")
-	elasticacheQA, elasticacheDatabaseQA, elasticacheQATags := makeElastiCacheCluster(t, "ec2", "us-east-1", "qa", withElastiCacheConfigurationEndpoint())
+	elasticacheProd, elasticacheDatabasesProd, elasticacheProdTags := makeElastiCacheCluster(t, "ec1", "us-east-1", "prod", mocks.WithElastiCacheReaderEndpoint)
+	elasticacheQA, elasticacheDatabasesQA, elasticacheQATags := makeElastiCacheCluster(t, "ec2", "us-east-1", "qa", mocks.WithElastiCacheConfigurationEndpoint)
 	elasticacheUnavailable, _, elasticacheUnavailableTags := makeElastiCacheCluster(t, "ec4", "us-east-1", "prod", func(cluster *elasticache.ReplicationGroup) {
 		cluster.Status = aws.String("deleting")
 	})
@@ -58,7 +57,7 @@ func TestElastiCacheFetcher(t *testing.T) {
 				},
 			},
 			inputMatchers: makeAWSMatchersForType(services.AWSMatcherElastiCache, "us-east-1", wildcardLabels),
-			wantDatabases: types.Databases{elasticacheDatabaseProd, elasticacheDatabaseQA},
+			wantDatabases: append(elasticacheDatabasesProd, elasticacheDatabasesQA...),
 		},
 		{
 			name: "fetch prod",
@@ -69,7 +68,7 @@ func TestElastiCacheFetcher(t *testing.T) {
 				},
 			},
 			inputMatchers: makeAWSMatchersForType(services.AWSMatcherElastiCache, "us-east-1", envProdLabels),
-			wantDatabases: types.Databases{elasticacheDatabaseProd},
+			wantDatabases: elasticacheDatabasesProd,
 		},
 		{
 			name: "skip unavailable",
@@ -80,7 +79,7 @@ func TestElastiCacheFetcher(t *testing.T) {
 				},
 			},
 			inputMatchers: makeAWSMatchersForType(services.AWSMatcherElastiCache, "us-east-1", wildcardLabels),
-			wantDatabases: types.Databases{elasticacheDatabaseProd},
+			wantDatabases: elasticacheDatabasesProd,
 		},
 		{
 			name: "skip unsupported",
@@ -91,31 +90,14 @@ func TestElastiCacheFetcher(t *testing.T) {
 				},
 			},
 			inputMatchers: makeAWSMatchersForType(services.AWSMatcherElastiCache, "us-east-1", wildcardLabels),
-			wantDatabases: types.Databases{elasticacheDatabaseProd},
+			wantDatabases: elasticacheDatabasesProd,
 		},
 	}
 	testAWSFetchers(t, tests...)
 }
 
-func makeElastiCacheCluster(t *testing.T, name, region, env string, opts ...func(*elasticache.ReplicationGroup)) (*elasticache.ReplicationGroup, types.Database, []*elasticache.Tag) {
-	cluster := &elasticache.ReplicationGroup{
-		ARN:                      aws.String(fmt.Sprintf("arn:aws:elasticache:%s:123456789012:replicationgroup:%s", region, name)),
-		ReplicationGroupId:       aws.String(name),
-		Status:                   aws.String("available"),
-		TransitEncryptionEnabled: aws.Bool(true),
-
-		// Default has one primary endpoint in the only node group.
-		NodeGroups: []*elasticache.NodeGroup{{
-			PrimaryEndpoint: &elasticache.Endpoint{
-				Address: aws.String("primary.localhost"),
-				Port:    aws.Int64(6379),
-			},
-		}},
-	}
-
-	for _, opt := range opts {
-		opt(cluster)
-	}
+func makeElastiCacheCluster(t *testing.T, name, region, env string, opts ...func(*elasticache.ReplicationGroup)) (*elasticache.ReplicationGroup, types.Databases, []*elasticache.Tag) {
+	cluster := mocks.ElastiCacheCluster(name, region, opts...)
 
 	tags := []*elasticache.Tag{{
 		Key:   aws.String("env"),
@@ -126,23 +108,10 @@ func makeElastiCacheCluster(t *testing.T, name, region, env string, opts ...func
 	if aws.BoolValue(cluster.ClusterEnabled) {
 		database, err := services.NewDatabaseFromElastiCacheConfigurationEndpoint(cluster, extraLabels)
 		require.NoError(t, err)
-		return cluster, database, tags
+		return cluster, types.Databases{database}, tags
 	}
 
 	databases, err := services.NewDatabasesFromElastiCacheNodeGroups(cluster, extraLabels)
 	require.NoError(t, err)
-	require.Len(t, databases, 1)
-	return cluster, databases[0], tags
-}
-
-// withElastiCacheConfigurationEndpoint returns an option function for
-// makeElastiCacheCluster to set a configuration endpoint.
-func withElastiCacheConfigurationEndpoint() func(*elasticache.ReplicationGroup) {
-	return func(cluster *elasticache.ReplicationGroup) {
-		cluster.ClusterEnabled = aws.Bool(true)
-		cluster.ConfigurationEndpoint = &elasticache.Endpoint{
-			Address: aws.String("configuration.localhost"),
-			Port:    aws.Int64(6379),
-		}
-	}
+	return cluster, databases, tags
 }
